requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
django-unfold==0.72.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==12.0.0
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.4
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
